[#124] Refactored entire client lib alternate, seperating outputting code from functional, and using ByteString
