+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/commandPaths/CommandExecutionPath.java
@@ -1,9 +1,6 @@
-package net.zscript.javaclient.core;
+package net.zscript.javaclient.commandPaths;
 
-import java.io.ByteArrayOutputStream;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Deque;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -11,12 +8,13 @@ import java.util.ListIterator;
 import java.util.Map;
 import java.util.Optional;
 
-import net.zscript.ascii.AsciiFrame;
+import net.zscript.javaclient.commandbuilder.ZscriptByteString;
 import net.zscript.javareceiver.tokenizer.TokenBuffer;
 import net.zscript.javareceiver.tokenizer.TokenBufferIterator;
 import net.zscript.javareceiver.tokenizer.Tokenizer;
 import net.zscript.model.ZscriptModel;
 import net.zscript.model.components.Zchars;
+import net.zscript.util.ByteString;
 
 public class CommandExecutionPath {
 
@@ -187,8 +185,8 @@ public class CommandExecutionPath {
         return cmds;
     }
 
-    public byte[] toSequence() {
-        ByteArrayOutputStream out = new ByteArrayOutputStream();
+    public ByteString toSequence() {
+        ZscriptByteString.ZscriptByteStringBuilder out = ZscriptByteString.builder();
 
         Command current = firstCommand;
         while (true) {
@@ -198,167 +196,30 @@ public class CommandExecutionPath {
                 if (link.getOnFail() == null) {
                     break;
                 } else {
-                    out.write(link.getFailSeparators()[0]);
+                    out.appendByte(link.getFailSeparators()[0]);
                     current = link.getOnFail();
                 }
             } else if (link.getSuccessSeparators()[0] == Zchars.Z_CLOSEPAREN) {
                 if (link.getOnFail() == null || link.getFailSeparators()[0] == Zchars.Z_CLOSEPAREN) {
-                    out.write(Zchars.Z_CLOSEPAREN);
+                    out.appendByte(Zchars.Z_CLOSEPAREN);
                     current = link.getOnSuccess();
                 } else {
-                    out.write(Zchars.Z_ORELSE);
+                    out.appendByte(Zchars.Z_ORELSE);
                     current = link.getOnFail();
                 }
             } else {
-                out.write(link.getSuccessSeparators()[0]);
+                out.appendByte(link.getSuccessSeparators()[0]);
                 current = link.getOnSuccess();
             }
         }
-        return out.toByteArray();
+        return out.build();
     }
 
-    public AsciiFrame graphPrint(CommandGrapher grapher) {
-        return graphPrint(grapher, CommandGraph.getDefaultSettings(), null);
+    public Command getFirstCommand() {
+        return firstCommand;
     }
 
-    public AsciiFrame graphPrint(CommandGrapher grapher, CommandGraph.GraphPrintSettings settings) {
-        return graphPrint(grapher, settings, null);
+    public ZscriptModel getModel() {
+        return model;
     }
-
-    public AsciiFrame graphPrint(CommandGrapher grapher, CommandGraph.GraphPrintSettings settings, ResponseExecutionPath resps) {
-        boolean skipImpossiblePaths = settings.skipImpossiblePaths();
-
-        CommandGrapher.CommandDepth        startDepth = new CommandGrapher.CommandDepth(0);
-        CommandGrapher.CommandGraphElement start      = new CommandGrapher.CommandGraphElement(firstCommand, startDepth);
-
-        Map<Command, CommandGrapher.CommandGraphElement> commands = new HashMap<>();
-
-        Deque<CommandGrapher.CommandGraphElement> openedTrees  = new ArrayDeque<>();
-        Deque<CommandGrapher.CommandGraphElement> workingTrees = new ArrayDeque<>();
-        List<CommandGrapher.CommandGraphElement>  elements     = new ArrayList<>();
-
-        commands.put(firstCommand, start);
-        workingTrees.push(start);
-        CommandGrapher.CommandDepth maxDepth = new CommandGrapher.CommandDepth(0);
-        while (!workingTrees.isEmpty()) {
-            CommandGrapher.CommandGraphElement current = workingTrees.peek();
-            Command.CommandEndLink             links   = current.getCommand().getEndLink();
-
-            CommandGrapher.CommandGraphElement latestOpenTree = openedTrees.peek();
-            if (latestOpenTree != null && links.getOnFail() != latestOpenTree.getCommand()) {
-                // if there are opened trees, and this command doesn't add to the top open tree,
-                //   check if it closes the top open tree
-                //   which we can do by iterating forward to see if the top open tree re-merges here
-                // this operation makes our graph drawing O(n^2)
-                Command tmp        = latestOpenTree.getCommand();
-                boolean mergesHere = false;
-                while (tmp != null) {
-                    if (tmp == current.getCommand()) {
-                        mergesHere = true;
-                        break;
-                    }
-                    tmp = tmp.getEndLink().getOnSuccess();
-                }
-                if (mergesHere) {
-                    openedTrees.pop();
-                    workingTrees.push(latestOpenTree);
-                    if (!openedTrees.isEmpty()) {
-                        openedTrees.peek().getDepth().depthGreaterThan(latestOpenTree.getDepth());
-                        maxDepth.depthGreaterThan(latestOpenTree.getDepth());
-                    }
-                    // if it closes the top open tree, process that tree first
-                    continue;
-                }
-            }
-            if (links.getOnFail() != null && (!skipImpossiblePaths || current.getCommand().canFail()) && (latestOpenTree == null
-                    || links.getOnFail() != latestOpenTree.getCommand())) {
-                // the command opens a new open tree
-                CommandGrapher.CommandGraphElement opened = new CommandGrapher.CommandGraphElement(links.getOnFail(), new CommandGrapher.CommandDepth(current.getDepth()));
-                maxDepth.depthGreaterThan(current.getDepth());
-                commands.put(opened.getCommand(), opened);
-                openedTrees.push(opened);
-                latestOpenTree = opened;
-            }
-
-            elements.add(workingTrees.pop()); // only put elements into the list when we're done processing them
-            if (links.getOnSuccess() != null && (!skipImpossiblePaths || current.getCommand().canSucceed())) {
-                if (workingTrees.isEmpty() || links.getOnSuccess() != workingTrees.peek().getCommand()) {
-                    // check we're not on a close parent, dropping out of failure
-                    CommandGrapher.CommandGraphElement next = new CommandGrapher.CommandGraphElement(links.getOnSuccess(), current.getDepth());
-                    commands.put(next.getCommand(), next);
-                    workingTrees.push(next);
-                }
-            } else if (latestOpenTree != null) {
-                openedTrees.pop();
-                workingTrees.push(latestOpenTree);
-                if (!openedTrees.isEmpty()) {
-                    openedTrees.peek().getDepth().depthGreaterThan(latestOpenTree.getDepth());
-                    maxDepth.depthGreaterThan(latestOpenTree.getDepth());
-                }
-            }
-        }
-        List<Response> responses;
-        List<Command>  toHighlight = List.of();
-        if (resps != null) {
-            responses = resps.getResponses();
-            toHighlight = this.compareResponses(resps);
-        } else {
-            responses = List.of();
-        }
-        if (skipImpossiblePaths) {
-            elements = skipEmpty(commands, elements);
-            toHighlight = new ArrayList<>(toHighlight);
-            Iterator<Response> respIter = responses.iterator();
-            for (Iterator<Command> iter = toHighlight.iterator(); iter.hasNext(); ) {
-
-                if (!elements.contains(commands.get(iter.next()))) {
-                    iter.remove();
-                    if (respIter.hasNext()) {
-                        respIter.next();
-                        respIter.remove();
-                    }
-                } else if (respIter.hasNext()) {
-                    respIter.next();
-                }
-            }
-            //then trim depths which aren't used...
-            int offset = 0;
-            for (int i = 0; i < maxDepth.getDepth() + 1; i++) {
-                boolean hasAtDepth = false;
-                for (CommandGrapher.CommandGraphElement element : elements) {
-                    if (element.getDepth().getDepth() == i) {
-                        hasAtDepth = true;
-                        element.getDepth().setDepth(i - offset);
-                    }
-                }
-                if (!hasAtDepth) {
-                    offset++;
-                }
-            }
-            maxDepth.setDepth(maxDepth.getDepth() - offset);
-        }
-        return grapher.graph(model, commands, elements, maxDepth, toHighlight, responses, settings);
-    }
-
-    private List<CommandGrapher.CommandGraphElement> skipEmpty(Map<Command, CommandGrapher.CommandGraphElement> commands,
-            List<CommandGrapher.CommandGraphElement> elements) {
-        List<CommandGrapher.CommandGraphElement> compactedElements = new ArrayList<>();
-        for (CommandGrapher.CommandGraphElement element : elements) {
-            if (element.getCommand().isEmpty()) {
-                CommandGrapher.CommandGraphElement tmp = element;
-                while (tmp.getCommand().isEmpty() && tmp.getCommand().getEndLink().getOnSuccess() != null) {
-                    tmp = commands.get(tmp.getCommand().getEndLink().getOnSuccess());
-                }
-                if ((!tmp.getCommand().isEmpty() || tmp.getCommand().getEndLink().getOnSuccess() != null) && element.getDepth().getDepth() != commands.get(
-                                element.getCommand().getEndLink().getOnSuccess())
-                        .getDepth().getDepth()) {
-                    compactedElements.add(element);
-                }
-            } else {
-                compactedElements.add(element);
-            }
-        }
-        return compactedElements;
-    }
-
 }
