+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/commandPaths/Response.java
@@ -1,9 +1,7 @@
-package net.zscript.javaclient.core;
+package net.zscript.javaclient.commandPaths;
 
-import java.io.ByteArrayOutputStream;
-import java.io.OutputStream;
-import java.nio.ByteBuffer;
-import java.nio.charset.StandardCharsets;
+
+import net.zscript.javaclient.commandbuilder.ZscriptByteString;
 
 public class Response {
     private final ZscriptFieldSet fieldSet;
@@ -24,15 +22,15 @@ public class Response {
         return next;
     }
 
-    public void toBytes(OutputStream out) {
+    public void toBytes(ZscriptByteString.ZscriptByteStringBuilder out) {
         fieldSet.toBytes(out);
     }
 
     @Override
     public String toString() {
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        toBytes(outputStream);
-        return StandardCharsets.UTF_8.decode(ByteBuffer.wrap(outputStream.toByteArray())).toString();
+        ZscriptByteString.ZscriptByteStringBuilder out = ZscriptByteString.builder();
+        toBytes(out);
+        return out.asString();
     }
 
     public ZscriptFieldSet getFields() {
