+++ clients/java-client-lib-alternate/client-core/src/main/java/net/zscript/javaclient/commandPaths/ResponseExecutionPath.java
@@ -1,4 +1,4 @@
-package net.zscript.javaclient.core;
+package net.zscript.javaclient.commandPaths;
 
 import java.util.ArrayList;
 import java.util.List;
