+++ clients/java-client-lib-alternate/client-core/src/test/java/net/zscript/javaclient/core/CompleteCommandGrapherTest.java
@@ -7,14 +7,15 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import net.zscript.ascii.AnsiCharacterStylePrinter;
-import net.zscript.javareceiver.semanticParser.SemanticAction;
+import net.zscript.javaclient.commandPaths.CommandExecutionPath;
+import net.zscript.javaclient.commandPrinting.CommandGraph;
+import net.zscript.javaclient.commandPrinting.StandardCommandGrapher;
+import net.zscript.javaclient.commandPrinting.VerbositySetting;
 import net.zscript.javareceiver.tokenizer.TokenExtendingBuffer;
 import net.zscript.javareceiver.tokenizer.Tokenizer;
 
@@ -33,7 +34,7 @@ public class CompleteCommandGrapherTest {
         CommandExecutionPath path = CommandExecutionPath.parse(bufferCmd.getTokenReader().getFirstReadToken());
 
         String ANSI_ANYTHING = "(\u001B[\\[\\d;]*m)*";
-        assertThat(path.graphPrint(new StandardCommandGrapher(), basicSettings).generateString(new AnsiCharacterStylePrinter()).replaceAll(ANSI_ANYTHING, "").lines()
+        assertThat(new StandardCommandGrapher().graph(path, null, basicSettings).generateString(new AnsiCharacterStylePrinter()).replaceAll(ANSI_ANYTHING, "").lines()
                 .map(s -> s.stripTrailing() + "\n").collect(
                         Collectors.joining())).
                 isEqualTo(output);
